django/requirements.txt
Django
gunicorn
python-decouple
requests
